feat: ajusta servicos para estudos(complete)

+++ src/useCases/(Study)/DeleteStudy/DeleteStudy_UseCase.ts
@@ -1,18 +1,26 @@
-import { NotFound } from "../../../repositories/IErrorRepository";
+import { IUploadImage } from "../../../providers/IUploadImage";
+import { NotFound, Unauthorized } from "../../../repositories/IErrorRepository";
 import { IStudyRepository } from "../../../repositories/IStudyRepository";
 import { IDeleteStudyDTO } from "./DeleteStudy_DTO";
 
 export class DeleteStudyUseCase {
-    constructor(
-        private studiesRepository: IStudyRepository
-    ) {}
-    
-    async execute(data: IDeleteStudyDTO): Promise<void> {
-        const studyExists = await this.studiesRepository.findById(data.id);
-        if (!studyExists) {
-            throw new NotFound("Estudo não encontrado no sistema");
-        }
-        
-        return await this.studiesRepository.deleteById(data.id);
+  constructor(
+    private studiesRepository: IStudyRepository,
+    private uploadThumbnail: IUploadImage
+  ) {}
+
+  async execute(data: IDeleteStudyDTO): Promise<void> {
+    const studyExists = await this.studiesRepository.findById(data.id);
+    if (!studyExists) {
+      throw new NotFound("Estudo não encontrado no sistema");
     }
+
+    if (studyExists.author !== data.authorId) {
+      throw new Unauthorized("Você não tem permissão para excluir este estudo");
+    }
+    await this.uploadThumbnail.destroy(studyExists.thumbnailId);
+    await this.studiesRepository.deleteById(data.id);
+
+    return;
+  }
 }
